+++ core/src/main/java/com/garden/back/global/jpa/JpaConfig.java
@@ -1,4 +1,3 @@
-
 package com.garden.back.global.jpa;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
